requirements.txt
fastapi
uvicorn
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-generativeai==0.8.5
google-api-core==2.25.1
python-dotenv
jinja2 # <--- ADD THIS LINE
python-multipart